Update User Guide to include edit command. Update input and output files to include edit command.
Add a new constructor for EditCommand to allow Person as an input parameter.
Reverted the change to private for the string message.
Change stringbuilder to concat + operator.

Add test cases for EditCommand with EditCommandTest class.

Change how the regex for targetIndex works. Now it only allows non-negative numeric numbers and the error message shows the correct one.